+++ spb/__init__.py
@@ -14,6 +14,7 @@ from spb.functions import (
     plot_polar,
     plot_geometry,
     plot_list,
+    plot3d_list,
     plot_piecewise
 )
 from spb.vectors import plot_vector
@@ -40,5 +41,5 @@ __all__ = [
     "plot3d_revolution", "plot_implicit", "plot_polar", "plot_geometry",
     "plot_list", "plot_piecewise", "plot_vector", "plot_complex",
     "plot_complex_list", "plot_real_imag", "plot_complex_vector", "plotgrid",
-    "MB", "BB", "PB", "KB", "MAB"
+    "MB", "BB", "PB", "KB", "MAB", "plot3d_list",
 ]